refactor: domain package
