+++ src/main/java/racingcar/MainApplication.java
@@ -1,6 +1,12 @@
 package racingcar;
 
 import java.util.List;
+import racingcar.domain.Car;
+import racingcar.domain.Names;
+import racingcar.domain.RaceEvaluator;
+import racingcar.domain.Racer;
+import racingcar.domain.Result;
+import racingcar.domain.Winner;
 import racingcar.view.InputView;
 import racingcar.view.ResultView;
