+++ src/main/java/racingcar/domain/Racer.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class Racer {
