+++ src/main/java/racingcar/domain/Result.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import java.util.HashMap;
 import java.util.Map;
